+++ src/css/index.css
@@ -36,3 +36,36 @@ li {
 a {
   text-decoration: none;
 }
+
+.todo-list {
+  background-color: white;
+  margin: 20px auto;
+  width: 50%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  border: 1px solid #efefef;
+  padding: 30px;
+}
+
+.todo-item {
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  padding:  30px 20px 0;
+  width: 70%;
+  border-bottom: 1px solid #cecece;
+  font-size: 15px;
+  font-weight: 200;
+  color: #333;
+}
+
+input[type=checkbox] {
+  margin-right: 10px;
+  padding: 20px;
+  font-size: 30px;
+}
+
+input[type=checkbox]:focus {
+  outline: none;
+}
